Increase PSBTMaxLen and MaxVaultsPerUser constants (#426)
* 🔧 chore(lib.rs): increase PSBTMaxLen and MaxVaultsPerUser constants
The PSBTMaxLen and MaxVaultsPerUser constants have been increased from 2048 to 4096 and from 10 to 100 respectively. This change allows for larger PSBTs and more vaults per user, which improves the flexibility and scalability of the system. The OutputDescriptorMaxLen constant has also been increased to 4096 to match the other changes.

* 🔖 chore(lib.rs): increase max proposals per vault to 100
The maximum number of proposals per vault has been increased from 5 to 100 to allow for more flexibility in the governance of the system. This change was made in the pallet_bitcoin_vaults::Config implementation in the runtime/src/lib.rs file.